fix authorization access in UserController

+++ src/Controller/UserController.php
@@ -8,15 +8,13 @@ use App\Repository\CommentRepository;
 use App\Repository\TrickRepository;
 use App\Helper\ImageFileDeletor;
 use App\Helper\UploaderHelper;
+use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
-/**
- * @IsGranted("access", subject="user", message="Access denied")
- */
 class UserController extends AbstractController
 {
     /**
@@ -30,6 +28,7 @@ class UserController extends AbstractController
     }
 
     /**
+     * @IsGranted("access", subject="user", message="Access denied")
      * @Route("/user/dashboard/{username}", name="user.dashboard")
      */
     public function dashboard(User $user, TrickRepository $trickRepository, CommentRepository $commentRepository)
@@ -46,6 +45,7 @@ class UserController extends AbstractController
     }
 
     /**
+     * @IsGranted("access", subject="user", message="Access denied")
      * @Route("/user/profile/{username}", name="user.profile")
      */
     public function profile(User $user)
@@ -57,6 +57,7 @@ class UserController extends AbstractController
     }
 
     /**
+     * @IsGranted("access", subject="user", message="Access denied")
      * @Route("/user/edit/{username}", name="user.edit")
      */
     public function edit(Request $request, User $user, UploaderHelper $uploaderHelper, ImageFileDeletor $imageFileDeletor)
@@ -74,8 +75,9 @@ class UserController extends AbstractController
             $this->em->persist($user);
             $this->em->flush();
 
-            $imageFileDeletor->deleteFile('user', $user->getId(), [$avatarName]);
-
+            if (!empty($avatar)) {
+                $imageFileDeletor->deleteFile('user', $user->getId(), [$avatarName]);
+            }
             $this->addFlash('success', 'Your profile has been updated !');
 
             return $this->redirectToRoute('user.profile', [
@@ -91,12 +93,13 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/users", name="users")
+     * @IsGranted("ROLE_ADMIN")
+     * @Route("/users", name="admin.users")
      */
-    public function users()
+    public function users(UserRepository $userRepository)
     {
-        $verifiedUsers = $this->userRepository->findBy(['activationToken' => null]);
-        $unverifiedUsers = $this->userRepository->findUnverified();
+        $verifiedUsers = $userRepository->findBy(['activationToken' => null]);
+        $unverifiedUsers = $userRepository->findUnverified();
 
         return $this->render('admin/users.html.twig', [
             'verifiedUsers' => $verifiedUsers,
